+++ NetworkXMCP/tools/__init__.py
@@ -5,19 +5,17 @@
 NetworkXを使用したグラフの操作ツールを提供するモジュール
 """
 
-from .network_tools import (
-    create_random_network,
-    parse_graphml_string,
-    convert_to_standard_graphml,
-    export_network_as_graphml,
-    get_network_info
-)
-
+from .graph_creation import create_random_network
+from .graphml_parser import parse_graphml_string, fix_graphml_structure
+from .graphml_converter import convert_to_standard_graphml, export_network_as_graphml
+from .network_analysis import get_network_info, calculate_centrality
 
 __all__ = [
     'create_random_network',
     'parse_graphml_string',
+    'fix_graphml_structure',
     'convert_to_standard_graphml',
     'export_network_as_graphml',
-    'get_network_info'
+    'get_network_info',
+    'calculate_centrality'
 ]
